Moved modal components to its own folder
