+++ Components/Modal/ModalContainer.tsx
@@ -1,7 +1,8 @@
 import React, { ReactNode } from "react";
 import styled from "styled-components/native";
-import { useTheme } from "../contexts/ThemeContext";
-import { colorsModel } from "../models/ColorsModel";
+import { useTheme } from "../../contexts/ThemeContext";
+import { colorsModel } from "../../models/ColorsModel";
+
 import ModalHeader from "./ModalHeader";
 
 interface Props {
